refactor(client-js): Restructured the modules into sub-categories
