+++ public/js/script.js
@@ -1,22 +1,18 @@
-    'use strict';
-
-    import * as header from './modules/header.js'
-    import {
-        sidebarNav
-    } from './modules/header-sidebar.js'
-    import Quiz from "./modules/quiz-setup.js";
-    import CustomSlider from "./modules/custom-slider.js"
+import * as header from './modules/header/header.js'
+import { sidebarNav } from './modules/header/header-sidebar.js'
+import Quiz from "./modules/quiz/quiz-setup.js";
+import CustomSlider from "./modules/custom-slider.js"
 
 
-    (() => {
-        header.headerBackgroundScroll()
-        sidebarNav()
+(() => {
+    header.headerBackgroundScroll()
+    sidebarNav()
 
-        const customSliders = document.getElementsByClassName("custom-slider")
+    const customSliders = document.getElementsByClassName("custom-slider")
 
-        for (let i = 0; i < customSliders.length; i++) {
-            new CustomSlider(customSliders[i], (origin, value) => {
-              origin.slider.pin.textContent = value;
-            })
-        }
-    })()
+    for (let i = 0; i < customSliders.length; i++) {
+      new CustomSlider(customSliders[i], (origin, value) => {
+        origin.slider.pin.textContent = value;
+      })
+    }
+})()
